organizing code a bit so it's not all in one file
